% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefit_model.R
\name{benefit_model}
\alias{benefit_model}
\title{Benefit model}
\arguments{
\item{.fleet}{The output of 'compliance_cost' function, with a compliant simulated fleet}

\item{.km_travelled}{The assumed km travelled for each vehicle segment and vehicle age}

\item{.fuel_prices}{Assumed future fuel prices (excluding taxes)}

\item{.electricity_prices}{Assumed future electricity prices}

\item{.energy_consumption}{Assumed energy consumption of electric vehicles}

\item{.energy_intensity}{Assumed energy intensity of the electricity grid}

\item{.gap}{The assumed 'gap' between test cycle and real world emissions for combustion vehicles}

\item{.fuel_scenario}{The assumed fuel price scenario}

\item{.electricity_scenario}{Assumed electricity price scenario}

\item{.run_to_year}{Year the model will run to}

\item{.passenger_diesel}{Proportion of disel passenger vehicles}

\item{.suv_diesel}{Proportion of suv diesel vehicles}

\item{.lcv_diesel}{Proportion of LCV diesel vehicles}

\item{.premium_95}{Proportion of vehicles requiring 95oct fuel}

\item{.premium_98}{Proportion of vehicles requiring 98oct fuel}
}
\value{
A \code{tibble} including the fuel, emissions, costs for each vehicle in the simulated fleet
over it' lifespan
}
\description{
Benefit model
}
