% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_costs.R
\name{compliance_costs}
\alias{compliance_costs}
\title{compliance_costs}
\usage{
compliance_costs(
  .fleet = fleet,
  .target = target,
  .cost_curves = cost_curves,
  .estimate = "icct_estimate",
  .suv_existing_tech = 20,
  .passenger_existing_tech = 13,
  .lcv_existing_tech = 15,
  .run_to_year = 2050
)
}
\arguments{
\item{.fleet}{The simulated fleet of new vehicle sales until 2050}

\item{.target}{The input target of a fleetwide efficiency standard to be evaluated or a business as usual trajectory}

\item{.cost_curves}{The assumed cost curves for al vehicle types}

\item{.estimate}{The cost curves estimate type}

\item{.suv_existing_tech}{The assumed existing technology for SUVs}

\item{.passenger_existing_tech}{The assumed existing technology for passenger vehicles}

\item{.lcv_existing_tech}{The assumed existing technology for LCVs}

\item{.run_to_year}{The year the function will run until}
}
\value{
A \code{tibble} with a 'compliant' vehicle fleet, where technology upgrades have been
applied to meet the specified target in each year
}
\description{
compliance_costs
}
\examples{



}
