% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_documentation.R
\docType{data}
\name{cost_curves}
\alias{cost_curves}
\title{Cost curves}
\format{
A \code{tibble} object with 8 variables:
\describe{
\item{\code{vehicle_group}}{The type of vehicle the upgrade applies to (passenger, suv, lcv)}
\item{\code{year}}{The year the upgrade applies to}
\item{\code{estimate}}{The upgrade estimate type - a central scenario (central), price parity I year ahead of of central (early_pp),
or one year behind central (late_pp)}
\item{\code{incr_cost}}{The incremental cost of applying the given upgrade (compared to the previous technology
package) for the given vehicle type, in a given year}
\item{\code{incr_reduction}}{The incremental emissions reduction (as a proportion of emissions stated between
0 and 1. 1 refers to a 100\% reduction in emissions.)}
\item{\code{type}}{The upgrade type - defined as either an upgrade to electric (ev) or
as an improvement to an ICE engine (ice)}
\item{\code{tech_pkg_no}}{The technology package to be applied. tech_pkg_no 100 refers to an electric vehicle.
Technology packages (excluding no 100) must be applied sequentially. Package 100 contains cost relative to base package
0.}
\item{\code{weighted_emissions}}{Vehicle emissions (historical and unused)}
}
}
\usage{
cost_curves
}
\description{
The costs and efficiency improvements associated with technology upgrades available to vehicles. Derived from US EPA cost curves used in the OMEGA model.
}
\keyword{datasets}
