% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{discount}
\alias{discount}
\title{Discount function}
\usage{
discount(data, rate)
}
\arguments{
\item{data}{Input data}

\item{rate}{Discount rate to be applied}
}
\value{
A discounted dataframe
}
\description{
Discount function
}
\examples{


}
