% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\docType{data}
\name{electricity_prices}
\alias{electricity_prices}
\title{Electricity prices}
\format{
A \code{tibble} object with 3 variables:
\describe{
\item{\code{scenario}}{The energy price scenario to be assumed. Options are "central",
"low_price", "high_price", and "off_peak"}
\item{\code{year}}{The year the cost applies to}
}
}
\usage{
electricity_prices
}
\description{
The assumed future electricity prices under different scenarios
}
\keyword{datasets}
