% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet_eff_sim.R
\name{fleet_eff_sim}
\alias{fleet_eff_sim}
\title{Fleet efficiency simulation}
\arguments{
\item{.fleet_creator}{Default is TRUE. Specifies whether the \code{fleet_creator} will be used to generate a simulated
fleet of vehicle sales. If FALSE, the fleet specified by \code{.input_fleet} will be used instead.}

\item{.in_cars}{Default is 100. The number of cars in the base year of the simulated fleet. To be used by \code{fleet_creator} to produce
a simulated fleet over all year. Although a larger simulated fleet sample (> 100) can be used, this is more time intensive and
is unlikely to yield significantly different results. For example, the BCR in simulations with \code{.in_cars = 100} and \code{.in cars = 400} differ by
~1\% overall.}

\item{.input_fleet}{The simulated 'fleet' to be run through the model as input by the user}

\item{.in_target_file}{Defaults to \code{targets_and_bau}. The file from where target and bau scenarios are drawn.}

\item{.in_target_scenario}{Defaults to "target_central". The assumed target trajectory scenario of a fleet standard from those contained in
\code{.in_target_file}}

\item{.in_bau_scenario}{Defaults to "bau". The assumed trajectory of the fleet standard from those contained in
\code{.in_target_file}}

\item{.in_cost_curves}{Defaults to "cost curves". The assumed costs of improving vehicle efficiency by year and vehicle type.}

\item{.in_cost_curve_estimate}{Defaults to "central". The estimate scenario used from the \code{cost_curves} scenarios. Options are "central",
"pp_late", "pp_early".}

\item{.in_suv_existing_tech}{Defaults to 20. The percentage assumed efficiency improvement of a base model new SUV vehicle when compared to
a I4 petrol engine (assumed as 2008 base model equivalent)}

\item{.in_passenger_existing_tech}{Defaults to 13 The percentage assumed efficiency improvement of a base model new passenger vehicle when compared to
a I4 petrol engine (assumed as 2008 base model equivalent)}

\item{.in_lcv_existing_tech}{Defaults to 13 The percentage assumed efficiency improvement of a base model new LCV vehicle when compared to
a I4 petrol engine (assumed as 2008 base model equivalent).}

\item{.in_run_to_year}{The year to whih the cost model will simulate.}

\item{.fleet}{The simulated fleet with technology upgrades applied to be run through the benefit model.}

\item{.in_km_travelled}{Defaults to \code{km_travelled}. The assumed distance travelled by each vehicle per year, depending on vehicle age and type.}

\item{.in_fuel_prices}{Defaults to \code{fuel_prices}. The assumed future price of fuel.}

\item{.in_electricity_prices}{Defaults to \code{electricity_prices}. the assumed future price of electricity.}

\item{.in_energy_consumption}{Defaults to \code{energy_consumption}. The assumed future energy consumption of electric vehicles.}

\item{.in_energy_intensity}{Defaults to \code{energy_intensity}. The assumed future energy intensity of the electricity grid.}

\item{.in_gap}{Defaults to 1.2. The assumed gap between tested and real world emissions. 1.2 is equivalent to 20\%}

\item{.in_fuel_scenario}{Defaults to "central". The assumed fuel price scenario from the \code{.in_fuel_prices} \code{tibble}}

\item{.in_electricity_scenario}{Defaults to "central". The assumed electricity price scenario from the \code{.in_electricity_prices} \code{tibble}}

\item{.in_benefit_run_to_year}{Defaults to 2060. The year to which the benefit model will run}

\item{.in_passenger_diesel}{Defaults to  0.02. The assumed proportion of passenger vehicles requiring diesel fuel.}

\item{.in_suv_diesel}{Defaults to 0.21. The assumed proportion of SUV vehicles requiring diesel fuel.}

\item{.in_lcv_diesel}{Defaults to 0.92. The assumed proportion of LCV vehicles requiring diesel fuel.}

\item{.in_premium_95}{Defaults to 0.15. The assumed proportion of ICE vehicles requiring premium 95ron petrol.}

\item{.in_premium_98}{Defaults to 0.05. The assumed proportion of ICE vehicles requiring premium 98ron petrol.}
}
\value{
A \code{tibble} containing the estimates vehicle costs (per vehicle),
fuel savings (per vehicle), emissions savings (total, Mt) and BCR under various discount rates.
}
\description{
This function runs the fleet efficiency simulation model over a simulated fleet of new vehicle sales between now
and 2050. It calculates the costs and benefits of a specified emissions target compared to the business as usual scenario.
}
