% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleeteffSim-package.R
\docType{package}
\name{fleeteffSim-package}
\alias{fleeteffSim}
\alias{fleeteffSim-package}
\title{fleeteffSim}
\description{
Simulate fleetwide efficiency standard
}
\author{
\strong{Maintainer}: Lachie Fox \email{lachlan.fox@grattaninstitute.edu.au}

}
\keyword{internal}
