% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_documentation.R
\docType{data}
\name{fuel_prices}
\alias{fuel_prices}
\title{Fuel prices}
\format{
A \code{tibble} object with 4 variables:
\describe{
\item{\code{scenario}}{The fuel price scenario to be assumed. Options are "central",
"low_price", "high_price"}
\item{\code{fuel_type}}{The type of fuel the cost applies to. Includes "diesel", "petrol_91",
"petrol_95" and "petrol_98"}
\item{\code{price}}{The assumed future price, excluding all taxes}
}
}
\usage{
fuel_prices
}
\description{
The assumed future prices of fuels
}
\keyword{datasets}
