% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_documentation.R
\docType{data}
\name{fuel_prices_no_tax}
\alias{fuel_prices_no_tax}
\title{Fuel prices without taxes}
\format{
A \code{tibble} object containing forecast fuel prices for use in the fleetEffSim model. Fuel prices are base prices only - they do not include fuel excise tax or GST. There are 4 variables:
\describe{
\item{\code{scenario}}{The fuel price scenario to be assumed. Options are "central",
"low_price", "high_price"}
\item{\code{fuel_type}}{The type of fuel the cost applies to. Includes "diesel", "petrol_91",
"petrol_95" and "petrol_98"}
\item{\code{price}}{The assumed future price, excluding all taxes}
}
}
\usage{
fuel_prices_no_tax
}
\description{
The assumed future prices of petrol and diesel fuels, excluding taxes.
}
\keyword{datasets}
