% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{generate_results}
\alias{generate_results}
\title{generate_results}
\usage{
generate_results(bau_benefits, target_benefits)
}
\arguments{
\item{bau_benefits}{the bau_benefits data}

\item{target_benefits}{the target benefits data}
}
\value{
summarised results
}
\description{
generate_results
}
\examples{


}
