% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_documentation.R
\docType{data}
\name{km_travelled}
\alias{km_travelled}
\title{KM travelled}
\format{
A \code{tibble} object with 3 variables:
\describe{
\item{\code{age}}{The age of the vehicle}
\item{\code{vehicle_type}}{The type of the vehicle (passenger, lcv, suv)}
\item{\code{km_travelled}}{The assumed dstance (km) travelled by the vehicle in a given year}
}
}
\usage{
km_travelled
}
\description{
The assumed distance driven in a given year, determined by vehicle type and age. Calculated from the Australian Bureau of Statistics (ABS) Motor Vehicle Use Survey (2016)
}
\keyword{datasets}
