% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_to_fuel.R
\name{petrol_co2_to_fuel}
\alias{petrol_co2_to_fuel}
\title{petrol_co2_to_fuel}
\usage{
petrol_co2_to_fuel(.co2)
}
\arguments{
\item{.co2}{The co2 intensity (g/km) to be converted}
}
\value{
The esimated fuel consumption of the vehicle (L/100km)
}
\description{
A function that converts vehicle carbon dioxide emissions (in grams/kilometre) to estimated fuel consumption (in L/100km) for petrol vehicles.
This is based on data provided by the National Transport Commission (NTC): https://www.ntc.gov.au/sites/default/files/assets/files/Carbon-dioxide-emissions-intensity-for-new-Australian-light-vehicles-2019.pdf
}
