% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_selector.R
\name{select_upgrade}
\alias{select_upgrade}
\title{upgrade_selector}
\usage{
select_upgrade(.this_year_fleet, .this_year_curves)
}
\arguments{
\item{.this_year_fleet}{A \code{tibble} containing the simulated fleet of vehicle sales for the iterated year}

\item{.this_year_curves}{A \code{tibble} containing the cost curves for vehicles sold in the iterated year}
}
\value{
A \code{tibble} with the selected most cost effective upgrade to apply to the fleet
}
\description{
upgrade_selector
}
\examples{



}
