% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{summarise_fes_results}
\alias{summarise_fes_results}
\title{Summarise FES results}
\usage{
summarise_fes_results(.data)
}
\arguments{
\item{.data}{}
}
\value{
A \code{tibble} with summarised results fora given discount rate scenario
}
\description{
Summarise FES results
}
