% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\docType{data}
\name{targets_and_bau}
\alias{targets_and_bau}
\title{Targets and Business as usual}
\format{
A \code{tibble} object with 3 variables:
\describe{
\item{\code{year}}{The year the emissions applies to}
\item{\code{target_type}}{The type of target or scenario. bau refers to the business as usual scenario. Other
scenarios ("central", "ambitious", "linear" refer to different targets that may be tested).}
}
}
\usage{
targets_and_bau
}
\description{
The business as usual emissions trajectory of new vehicle sales, and potential fleetwide
emissions standards that may be implemented
}
\keyword{datasets}
