% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_existing_tech.R
\name{add_existing_technology}
\alias{add_existing_technology}
\title{Add existing technology}
\arguments{
\item{.type}{The type of vehicle (ie. passenger, suv) to be applied to}

\item{.existing_tech}{The assumed level of existing technology, as a percentage reduction of co2 since 2008 average}

\item{.estimate}{The cost curve estimate type}

\item{.cost_curves}{The cost curves used in the overall model run}
}
\value{
A \code{tibble} containing updated cost curves, where the costs of assumed 'existing' technology is 0
}
\description{
Add existing technology
}
