% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\docType{data}
\name{cost_curves}
\alias{cost_curves}
\title{Cost curves}
\format{

}
\usage{
cost_curves
}
\description{
The costs and efficiency improvements associated with technology upgrades available to vehicles.
Derived from US EPA cost curves as part of the OMEGA model.
}
\keyword{datasets}
