% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_documentation.R
\docType{data}
\name{fleet}
\alias{fleet}
\title{Fleet}
\format{
A \code{tibble} object with 8 variables:
\describe{
\item{\code{vehicle_group}}{The type of vehicle}
\item{\code{year}}{The year vehicle is 'sold' in}
\item{\code{id}}{The ID for the vehicle in the given sales year}
\item{\code{base_emissions}}{The assumed emissions of the vehicle if it is a 'base' vehicle (
considered to be a 4-cylinder petrol engine of similar spec to 2008 vehicle sales)}
\item{\code{current_emissions}}{The assumed emissions of the vehicle after upgrades are
applied}
\item{\code{electric_applied}}{A logical indicating if the vehicle is electric or not}
\item{\code{tech_pkg_applied}}{A value representing the highest level technology package number the vehicle
has applied. This is it's position on the cost curve}
\item{\code{cost}}{The total cost of upgrades applied to the vehicle}
}
}
\usage{
fleet
}
\description{
A simulated fleet of new vehicle sales from 2021 to 2035, as produced by 'fleet_creator()'. Assumes a default of 100 cars sold in the base year. Each car therefore represents 1\% of new Australian car sales.
}
\keyword{datasets}
