% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet_creator.R
\name{fleet_creator}
\alias{fleet_creator}
\title{Fleet creator}
\arguments{
\item{.i_cars}{The number of cars sold in the base year (2021) in the simulated fleet}

\item{.i_passenger_share}{The share of passenger vehicles in the base year}

\item{.i_suv_share}{The share of SUVs in the base year}

\item{.i_lcv_share}{The share of Light Commercial Vehicles in the base year}

\item{.passenger_growth}{The growth rate of passenger vehicle sales per year}

\item{.suv_growth}{The growth rate of SUV sales per year}

\item{.lcv_growth}{The growth rate of LCV sales per year}

\item{.passenger_co2}{The assumed average co2 emissions of passenger vehicles in base year}

\item{.suv_co2}{The assumed average co2 emissions of SUV vehicles in base year}

\item{.lcv_co2}{The assumed average co2 emissions of LCV vehicles in base year}

\item{.digits}{Used for rounding}
}
\value{
a \code{tibble} with each row representing a vehicle.
}
\description{
Creates a fleet of simulated vehicle sales until 2050, characterised by vehicle
segment and emissions.
}
