% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{generate_results}
\alias{generate_results}
\title{Generate results}
\arguments{
\item{bau_benefits}{The bau scenario output of the benefit_model()}

\item{target_benefits}{The target scenario output of the benefit_model()}

\item{cars}{The number of cars in the base year of the simulation run. By default, fleet_creator() sets 100 cars.}
}
\value{
summarised results
}
\description{
Generate summarised and discounted results from a business as usual and target scenario output of the benefit_model().
}
