% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{generate_results}
\alias{generate_results}
\title{Generate results}
\arguments{
\item{bau_benefits}{the bau_benefits data}

\item{target_benefits}{the target benefits data}

\item{cars}{the number of cars in the base year of the simulation run}
}
\value{
summarised results
}
\description{
Generate results
}
