% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_to_fuel.R
\name{petrol_co2_to_fuel}
\alias{petrol_co2_to_fuel}
\title{petrol_co2_to_fuel}
\usage{
petrol_co2_to_fuel(.co2)
}
\arguments{
\item{.co2}{The co2 intensity (g/km) to be converted}
}
\value{
A value of the fuel consumed
}
\description{
petrol_co2_to_fuel
}
\examples{


}
