% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{summarise_fes_results}
\alias{summarise_fes_results}
\title{Summarise fleetEffSim results from the benefit_model() function.}
\usage{
summarise_fes_results(.data)
}
\arguments{
\item{.data}{The input data to be summarised}
}
\value{
A \code{tibble} with summarised results fora given discount rate scenario
}
\description{
Summarise fleetEffSim results from the benefit_model() function.
}
